requirements.txt
gradio
pandas
numpy
matplotlib
seaborn
scikit-learn
lightgbm
xgboost
joblib
requests

google-generativeai
python-dotenv

openai
